fixed baseDamage, set to double

+++ src/modules/cannon/entities/cannon.entity.ts
@@ -15,7 +15,7 @@ export class Cannon extends Model {
   @Column({ type: "int" })
   imageId: number;
 
-  @Column({ type: "int" })
+  @Column({ type: "double" })
   baseDamage: number = 0;
 
   @Column({ type: "text" })
